requirements.txt
# ctrl-arm emg+imu gesture recognition

# hardware interface
pyserial>=3.5
numpy>=1.20.0
pandas>=1.5.0
PyYAML>=6.0

# ml pipeline
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
scikit-learn>=1.2.0
scipy>=1.9.0

# voice recognition
SpeechRecognition>=3.8.0
pyaudio>=0.2.11
pocketsphinx>=5.0.0

# visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# signal processing
pywavelets>=1.3.0
librosa>=0.9.0

# model deployment
onnx>=1.12.0
onnxruntime>=1.12.0

# adk
google-adk>=1.15.1

# voice console runner deps
openai-whisper>=2024.04.12
sounddevice>=0.4.6
python-dotenv>=1.0.1
google-genai>=0.7.0
